src/socket-worker/requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
ffmpeg-python==0.2.0
future==0.18.2
idna==2.10
numpy==1.20.1
opencv-python==4.5.1.48
pygame==2.0.1
python-engineio==4.0.0
python-socketio==5.0.4
requests==2.25.1
schedule==1.0.0
six==1.15.0
urllib3==1.26.3
websocket-client==0.57.0
